+++ backend/apps/cards/views.py
@@ -1,9 +1,8 @@
 from rest_framework import mixins
 from rest_framework.viewsets import GenericViewSet
 
-from apps.accounts.models import CustomUser
 from apps.cards.models import Card, Deck, Leader
-from apps.cards.serializers import CardSerializer, DeckSerializer, LeaderSerializer, UserCardsSerializer
+from apps.cards.serializers import CardSerializer, DeckSerializer, LeaderSerializer
 
 
 class CardViewSet(GenericViewSet, mixins.ListModelMixin):
@@ -32,9 +31,3 @@ class DeckViewSet(GenericViewSet,
 class LeaderViewSet(GenericViewSet, mixins.ListModelMixin):
     queryset = Leader.objects.select_related("ability", "faction").all()
     serializer_class = LeaderSerializer
-
-
-class UserDatabaseViewSet(GenericViewSet, mixins.RetrieveModelMixin):
-    """user_database, id: user_id"""
-    queryset = CustomUser.objects.all()
-    serializer_class = UserCardsSerializer
