+++ backend/apps/user_database/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class UserDatabaseConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.user_database'
